+++ packages/sdk/src/util/api-model.ts
@@ -1,8 +1,8 @@
 // For VC SDK
-export * as VCApiTypes from '@meeco/vc-api-sdk/lib/types/models';
+export * as VCApiTypes from '@meeco/vc-api-sdk/lib/types';
 // For Identity Network SDK
-export * as IdentityNetworkApiTypes from '@meeco/identity-network-api-sdk/lib/types/models';
+export * as IdentityNetworkApiTypes from '@meeco/identity-network-api-sdk/lib/types';
 // For Keystore SDK
-export * as KeystoreApiTypes from '@meeco/keystore-api-sdk/lib/types/models';
+export * as KeystoreApiTypes from '@meeco/keystore-api-sdk/lib/types';
 // For Vault SDK
-export * as VaultApiTypes from '@meeco/vault-api-sdk/lib/types/models';
+export * as VaultApiTypes from '@meeco/vault-api-sdk/lib/types';